% Auto-generated: do not edit by hand
\name{''MusicComponent}

\alias{''MusicComponent}

\title{MusicComponent component}

\description{

}

\usage{
''MusicComponent(id=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

