% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{allEdges}
\alias{allEdges}
\title{All edges of a graph}
\usage{
allEdges(amat)
}
\arguments{
\item{amat}{A square Boolean matrix, with dimnames, the adjacency matrix of
a graph.}
}
\value{
A matrix with two columns. Each row of the matrix is a pair of
indices indicating an edge of the graph. If the graph is undirected, then
only one of the pairs \eqn{(i,j), (j,i)} is reported.
}
\description{
Finds the set of edges of a graph. That is the set of undirected edges if
the graph is undirected and the set of arrows if the graph is directed.
}
\examples{
## A UG graph
allEdges(UG(~ y * v * k + v * k * d + y * d))

## A DAG
allEdges(DAG(u ~ h + o + p, h ~ o, o ~ p))

}
\seealso{
\code{\link[=cycleMatrix]{cycleMatrix()}}
}
\author{
Giovanni M. Marchetti
}
\keyword{graphs}
\keyword{models}
\keyword{multivariate}
