% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_annotators}
\alias{cache_annotators}
\title{Available WFDB annotator sidecars}
\usage{
cache_annotators(cache)
}
\arguments{
\item{cache}{A \code{StudyCache}.}
}
\value{
Character vector of annotator extensions.
}
\description{
Available WFDB annotator sidecars
}
