% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{correlations}
\alias{correlations}
\title{Marginal and partial correlations}
\usage{
correlations(x)
}
\arguments{
\item{x}{a square symmetric matrix, a covariance matrix, or a data.frame for
n observations and p variables.}
}
\value{
a square correlation matrix with marginal correlations (lower
triangle) and partial correlations (upper triangle).
}
\description{
Computes a correlation matrix with ones along the diagonal, marginal
correlations in the lower triangle and partial correlations given all
remaining variables in the upper triangle.
}
\examples{

## See Table 6.1 in Cox & Wermuth (1996)
data(glucose)
correlations(glucose)

}
\references{
Cox, D. R. & Wermuth, N. (1996). \emph{Multivariate
dependencies}. London: Chapman & Hall.
}
\seealso{
\code{\link{parcor}}, \code{\link{cor}}
}
\author{
Giovanni M. Marchetti
}
\keyword{array}
\keyword{graphs}
\keyword{models}
\keyword{multivariate}
