% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_funs.R
\name{fitUnivariateReg}
\alias{fitUnivariateReg}
\title{Fit linear recursive regressions with independent residuals
  ...,  a list of models
  data, a data frame.}
\usage{
fitUnivariateReg(..., data)
}
\description{
Fit linear recursive regressions with independent residuals
  ...,  a list of models
  data, a data frame.
}
