% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{fundCycles}
\alias{fundCycles}
\title{Fundamental cycles}
\usage{
fundCycles(amat)
}
\arguments{
\item{amat}{A symmetric matrix with dimnames denoting
the adjacency matrix of the undirected graph. The graph must
be connected, otherwise the function returns \code{NULL}.}
}
\value{
A list of matrices with two columns.
Every component of the list is  associated to a cycle.
The cycle is described by a \eqn{k \times 2}{k x 2}
matrix whose rows are the edges of the cycle.
If there is no cycle, the function returns \code{NULL}.
}
\description{
Finds the list of fundamental cycles of a connected
undirected graph.
}
\details{
All the cycles in an UG can be obtained from combination
(ring sum) of the set  of fundamental cycles.

This function is used by \code{\link[=cycleMatrix]{cycleMatrix()}} and
\code{\link[=isGident]{isGident()}}.
}
\examples{
## Three fundamental cycles
fundCycles(UG(~ a * b * d + d * e + e * a * f))

}
\references{
Thulasiraman, K. & Swamy, M.N.S. (1992).
\emph{Graphs: theory and algorithms}. New York: Wiley.
}
\seealso{
\code{\link[=UG]{UG()}}, \code{\link[=findPath]{findPath()}}, \code{\link[=cycleMatrix]{cycleMatrix()}}
\code{\link[=isGident]{isGident()}}, \code{\link[=bfsearch]{bfsearch()}}
}
\author{
Giovanni M. Marchetti
}
\keyword{graphs}
\keyword{models}
\keyword{multivariate}
