% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-uplot.R
\name{ggm_plotOutput}
\alias{ggm_plotOutput}
\title{Shiny output binding for a ggm plot}
\usage{
ggm_plotOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{Output variable to read from.}

\item{width, height}{Valid CSS dimensions.}
}
\value{
A Shiny widget output element.
}
\description{
Shiny output binding for a ggm plot
}
