% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracing-svg.R
\name{gram_tracingOutput}
\alias{gram_tracingOutput}
\title{Shiny output binding for a tracing}
\usage{
gram_tracingOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{Output variable to read from.}

\item{width, height}{Valid CSS dimensions.}
}
\value{
A Shiny widget output element.
}
\description{
Shiny output binding for a tracing
}
\seealso{
Other widgets:
\code{\link[=gram_plot]{gram_plot()}},
\code{\link[=gram_plotOutput]{gram_plotOutput()}},
\code{\link[=render_gram_plot]{render_gram_plot()}},
\code{\link[=render_gram_tracing]{render_gram_tracing()}}
}
\concept{widgets}
\keyword{internal}
