% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{In}
\alias{In}
\title{Indicator matrix}
\usage{
In(A)
}
\arguments{
\item{A}{a matrix.}
}
\value{
a matrix of the same dimensions as \code{A}.
}
\description{
Finds the indicator matrix of the zeros of a matrix.
}
\details{
The indicator matrix is a matrix of zeros and ones which has a zero element
iff the corresponding element of \code{A} is (exactly) zero.
}
\examples{

# A simple way to find the overall induced concentration graph
# The DAG on p. 198 of Cox & Wermuth (1996)
amat <- DAG(y1 ~ y2 + y3, y3 ~ y5, y4 ~ y5)
A <- edgematrix(amat)
In(crossprod(A))

}
\references{
Wermuth, N. & Cox, D.R. (2004). Joint response graphs and
separation induced by triangular systems. \emph{J.R. Statist. Soc. B},
66, Part 3, 687-717.
}
\seealso{
\code{\link[=DAG]{DAG()}}, \code{\link[=inducedCovGraph]{inducedCovGraph()}}, \code{\link[=inducedConGraph]{inducedConGraph()}},
\code{\link[=In]{In()}}.
}
\author{
Giovanni M. Marchetti
}
\keyword{algebra}
\keyword{array}
\keyword{graphs}
\keyword{multivariate}
