% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{likGau}
\alias{likGau}
\title{Deviance of a Gaussian graphical model}
\usage{
likGau(K, S, n, k)
}
\arguments{
\item{K}{A square concentration matrix (inverse covariance matrix).}

\item{S}{A square sample covariance matrix.}

\item{n}{An integer indicating the sample size.}

\item{k}{An integer indicating the number of variables
(number of rows of \code{S}).}
}
\value{
A numeric value representing the deviance of the Gaussian model.
}
\description{
Computes the deviance of a Gaussian model given the concentration matrix,
the sample covariance matrix, the sample size, and the number of variables.
}
\seealso{
\code{\link{log}}, \code{\link{det}}, \code{\link{diag}}
}
\author{
Kayvan Sadeghi, Giovanni M. Marchetti
}
