% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{MarkEqMag}
\alias{MarkEqMag}
\title{Markov equivalence of maximal ancestral graphs}
\usage{
MarkEqMag(amat, bmat)
}
\arguments{
\item{amat}{An adjacency matrix, or a graph object. This can be:
\itemize{
\item a \code{graphNEL} object,
\item an \code{\link[igraph:igraph]{igraph::igraph()}} object, or
\item a character vector of length \eqn{3e}{3e}, where \eqn{e}
is the number of edges. If it is a vector, it must
be a sequence of triples (\code{type}, \code{node1label}, \code{node2label}).

The type of edge can be:
\itemize{
\item \code{"a"} (arrow from \code{node1} to \code{node2}),
\item \code{"b"} (bi-directed arc),
\item \code{"l"} (undirected line), or
\item \code{"*"} (for an isolated node with \code{node1 == node2}).
}
}}

\item{bmat}{The same as \code{amat}.}
}
\value{
A character string: either \code{"Markov Equivalent"} or \code{"NOT Markov Equivalent"}.
}
\description{
\code{MarkEqMag} determines whether two maximal ancestral graphs (MAGs)
are Markov equivalent. The function checks whether the two graphs
have the same skeleton and colliders with order.
}
\details{
@details Support for \code{graphNEL} objects requires the \code{graph} package
from Bioconductor, which is a suggested dependency.
If the package is missing, passing a \code{graphNEL} object will
trigger an informative error.
}
\examples{
H1 <- matrix(c(
  0, 100, 0, 0,
  100, 0, 100, 0,
  0, 100, 0, 100,
  0, 1, 100, 0
), 4, 4)

H2 <- matrix(c(
  0, 0, 0, 0, 
  1, 0, 100, 0, 
  0, 100, 0, 100, 
  0, 1, 100, 0
), 4, 4)

H3 <- matrix(c(
  0, 0, 0, 0, 
  1, 0, 0, 0, 
  0, 1, 0, 100, 
  0, 1, 100, 0
), 4, 4)

MarkEqMag(H1, H2)
MarkEqMag(H1, H3)
MarkEqMag(H2, H3)

}
\references{
Ali, R.A., Richardson, T.S. and Spirtes, P. (2009)
Markov equivalence for ancestral graphs. \emph{Annals of Statistics},
37(5B), 2808-2837.
}
\seealso{
\code{\link[=MarkEqRcg]{MarkEqRcg()}}, \code{\link[=msep]{msep()}}
}
\author{
Kayvan Sadeghi
}
\keyword{MAG}
\keyword{Markov-equivalence}
\keyword{maximal}
