% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{pcor_test}
\alias{pcor_test}
\alias{pcor.test}
\title{Test for zero partial association}
\usage{
pcor_test(r, q, n)

pcor.test(r, q, n)
}
\arguments{
\item{r}{a partial correlation coefficient, computed by \code{pcor()}.}

\item{q}{the number of variables in the conditioning set.}

\item{n}{integer \eqn{> 0}, the sample size.}
}
\value{
\itemize{
\item \code{tval} The Student's t-test statistic.
\item \code{df} The degrees of freedom
\item \code{pvalue} The p-value, assuming a two-sided alternative.
}
}
\description{
Test for conditional independence between two variables,
given the other ones, assuming a multivariate normal distribution.
}
\details{
\code{pcor.test()} is deprecated and will be removed in future
versions to avoid conflicts with R's S3 class system.
Use \code{pcor_test()}.
}
\examples{

## Are 2,3 independent given 1?
data(marks)
pcor_test(pcor(c(2, 3, 1), var(marks)), 1, n = 88)

}
\seealso{
\code{pcor}
}
\author{
Giovanni M. Marchetti
}
\keyword{multivariate}
\keyword{test}
