% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-uplot.R
\name{render_ggm_plot}
\alias{render_ggm_plot}
\title{Shiny render function for a ggm plot}
\usage{
render_ggm_plot(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{Expression that produces a \code{\link[=ggm_plot]{ggm_plot()}} widget.}

\item{env}{Environment in which to evaluate \code{expr}.}

\item{quoted}{Whether \code{expr} is quoted.}
}
\value{
A Shiny render function.
}
\description{
Shiny render function for a ggm plot
}
