% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-svg.R
\name{render_ggm_tracing}
\alias{render_ggm_tracing}
\title{Shiny render function for a tracing}
\usage{
render_ggm_tracing(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{Expression that produces a \code{\link[=ggm_tracing]{ggm_tracing()}} widget.}

\item{env}{Environment in which to evaluate \code{expr}.}

\item{quoted}{Whether \code{expr} is quoted.}
}
\value{
A Shiny render function.
}
\description{
Shiny render function for a tracing
}
