% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-uplot.R
\name{render_gram_plot}
\alias{render_gram_plot}
\title{Shiny render function for a gram plot}
\usage{
render_gram_plot(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{Expression that produces a \code{\link[=gram_plot]{gram_plot()}} widget.}

\item{env}{Environment in which to evaluate \code{expr}.}

\item{quoted}{Whether \code{expr} is quoted.}
}
\value{
A Shiny render function.
}
\description{
Shiny render function for a gram plot
}
