% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{SG}
\alias{SG}
\title{Summary graph}
\usage{
SG(
  amat,
  M = c(),
  C = c(),
  showmat = TRUE,
  plot = FALSE,
  plotfun = plotGraph,
  ...
)
}
\arguments{
\item{amat}{An adjacency matrix, or a graph that can be a
\code{graphNEL} or an \code{\link[igraph:igraph]{igraph::igraph()}} object or a vector of
length \eqn{3e}, where \eqn{e} is the number of edges of the graph,
that is a sequence of triples \verb{(type, node1label, node2label)}.
The type of edge can be
\itemize{
\item \code{"a"} (arrows from \code{node1} to \code{node2})
\item \code{"b"} (arcs), and
\item \code{"l"} (lines).
}}

\item{M}{A subset of the node set of \code{amat} that is going to be
marginalised over.}

\item{C}{Another disjoint subset of the node set of \code{amat} that is
going to be conditioned on.}

\item{showmat}{A logical value. \code{TRUE} (by default) to print the
generated matrix.}

\item{plot}{A logical value, \code{FALSE} (by default). \code{TRUE} to plot
the generated graph.}

\item{plotfun}{Function to plot the graph when \code{plot = TRUE}.
Can be \code{plotGraph} (the default) or \code{drawGraph}.}

\item{...}{Further arguments passed to \code{plotfun}.}
}
\value{
A matrix that consists 4 different integers as an
\eqn{ij}-element:
\itemize{
\item 0 for a missing edge between \eqn{i} and \eqn{j},
\item 1 for an arrow from \eqn{i} to \eqn{j},
\item 10 for a full line between \eqn{i} and \eqn{j}, and
\item 100 for a bi-directed arrow between \eqn{i} and \eqn{j}.
}

These numbers are added to be associated with multiple edges
of different types. The matrix is symmetric w.r.t full lines
and bi-directed arrows.
}
\description{
\code{SG} generates and plots summary graphs after marginalization and
conditioning.
}
\examples{
# The adjacency matrix of a DAG
ex <- matrix(c(
  0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 
  0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
  1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
  0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
  0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
  0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
  0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
  0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
  0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0,
  0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0,
  0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0,
  1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
  0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0,
  0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
  1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0,
  0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0
), 16, 16, byrow = TRUE)
M <- c(3, 5, 6, 15, 16)
C <- c(4, 7)
SG(ex, M, C, plot = TRUE)
SG(ex, M, C, plot = TRUE, plotfun = drawGraph, adjust = FALSE)

}
\references{
Sadeghi, K. (2013). Stable mixed graphs. \emph{Bernoulli}, 19(5B),
2330–2358.

Wermuth, N. (2011). Probability distributions with summary graph
structure. \emph{Bernoulli}, 17(3),845-879.
}
\seealso{
\code{\link[=AG]{AG()}}, \code{\link[=MSG]{MSG()}}, \code{\link[=RG]{RG()}}
}
\author{
Kayvan Sadeghi
}
\keyword{DAG}
\keyword{conditioning}
\keyword{marginalization}
\keyword{summary-graph}
