% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{SPl}
\alias{SPl}
\title{Find element positions in a sorted vector}
\usage{
SPl(a, b)
}
\arguments{
\item{a}{A vector (character or numeric) to be sorted and analyzed.}

\item{b}{A vector containing the elements to search for.}
}
\value{
An integer vector representing the indices of the matches within the sorted vector.
}
\description{
Sorts the input vector \code{a} and returns the indices of the
positions where the elements contained in \code{b} are found.
}
\seealso{
\code{\link{sort}}, \code{\link[base]{\%in\%}}, \code{\link{which}}
}
\author{
Kayvan Sadeghi, Giovanni M. Marchetti
}
