\name{Utility Functions}
\alias{rem}
\alias{SPl}
\alias{RR}
\alias{likGau}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Utility functions 
}
\description{
Functions used internally.
}
%- maybe also 'usage' for other objects documented here.

\author{
Kayvan Sadeghi, Giovanni M. Marchetti
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{unique},\code{setdiff}, \code{is.element}
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{utility}

