% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggm-package.R
\docType{data}
\name{anger}
\alias{anger}
\title{Anger data}
\format{
A covariance matrix for 4 variables measured on 684 female students.
\describe{ \item{X}{anxiety state} \item{Y}{anger state} \item{Z}{anxiety
trait} \item{U}{anger trait} }
}
\usage{
anger
}
\description{
Anger data
}
\details{
Trait variables are viewed as stable personality characteristics, and state
variables denote behaviour in specific situations.  See Cox and Wermuth
(1996).
}
\examples{
 
# Fit a chordless 4-cycle model 
data(anger) 
G = UG(~ Y*X + X*Z + Z*U + U*Y)
fitConGraph(G,anger, 684) 

}
\references{
Cox, D. R. and Wermuth, N. (1996). \emph{Multivariate
dependencies}. London: Chapman and Hall.

Cox, D.R. and Wermuth, N. (1990). \emph{An approximation to maximum
likelihood estimates in reduced models}. 77(4), 747-761.
}
\keyword{datasets}
