% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{binve}
\alias{binve}
\title{Inverts a marginal log-linear parametrization}
\usage{
binve(eta, C, M, G, maxit = 500, print = FALSE, tol = 1e-10)
}
\arguments{
\item{eta}{a vector of dimension \code{t-1} where \code{t} is the number of cells of a contingency table.}

\item{C}{A contrast matrix.}

\item{M}{A marginalization matrix.}

\item{G}{G is the model matrix of the loglinear parameterization with no  constant term.}

\item{maxit}{an integer, specifying the maximum number of iterations. Default 500.}

\item{print}{a logical value: if \code{TRUE}, prints the criterion after each cycle.}

\item{tol}{A small value specifying the tolerance for the convergence criterion. Default: \code{1e-10}.}
}
\value{
A vector of probabilities \code{p}.
}
\description{
Inverts a marginal log-linear parametrization.
}
\details{
A marginal log-linear link is defined by \eqn{\eta = C (M \log p)}. See Bartolucci et al. (2007).
}
\note{
From a Matlab function by A. Forcina, University of Perugia, Italy.
}
\references{
Bartolucci, F., Colombi, R. and Forcina, A. (2007). An extended
class of marginal link functions for modelling contingency tables by
equality and inequality constraints. Statist. Sinica 17, 691-711.
}
\seealso{
\code{\link{mat.mlogit}}
}
\author{
Antonio Forcina, Giovanni M. Marchetti
}
\keyword{data}
\keyword{discrete}
\keyword{log-linear}
\keyword{marginal}
\keyword{models}
