% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{blkdiag}
\alias{blkdiag}
\title{Block diagonal matrix}
\usage{
blkdiag(...)
}
\arguments{
\item{\dots}{Variable number of matrices \code{M1, M2, ...}.}
}
\value{
A block diagonal matrix \code{diag(M1, M2, ...)}.
}
\description{
Block diagonal concatenation of input arguments.
}
\examples{

X <- c(1,1,2,2); Z <- c(10, 20, 30, 40); A <- factor(c(1,2,2,2))
blkdiag(model.matrix(~X+Z), model.matrix(~A))

}
\seealso{
\code{\link{diag}}
}
\author{
Giovanni M. Marchetti
}
\keyword{matrix}
