% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{blodiag}
\alias{blodiag}
\title{Block diagonal matrix}
\usage{
blodiag(x, blo)
}
\arguments{
\item{x}{A vector of length \code{n}.}

\item{blo}{A vector of positive integers such that \code{sum(blo) == n}.}
}
\value{
A block-diagonal matrix with as many row as elements of \code{blo}
and \code{n} columns. The vector \code{x} is split into \code{length(blo)}
sub-vectors and these are the blocks of the resulting matrix.
}
\description{
Split a vector x into a block diagonal matrix.
}
\examples{

blodiag(1:10, blo = c(2, 3, 5)) 
blodiag(1:10, blo = c(3,4,0,1))

}
\seealso{
\code{\link{blkdiag}}, \code{\link{diag}}
}
\author{
Giovanni M. Marchetti
}
\keyword{matrix}
