% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_paths}
\alias{cache_paths}
\title{All paths in a study group}
\usage{
cache_paths(cache)
}
\arguments{
\item{cache}{A \code{StudyCache}.}
}
\value{
Named character vector of canonical and cache paths.
}
\description{
All paths in a study group
}
