% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{cmpGraph}
\alias{cmpGraph}
\title{The complementary graph}
\usage{
cmpGraph(amat, value = 1)
}
\arguments{
\item{amat}{The adjacency matrix of an undirected graph.}

\item{value}{the value 1 or 10 indicating the edges present. The default is
\code{value = 1}.}
}
\value{
The edge matrix of the complementary graph.
}
\description{
Finds the complementary graph of an undirected graph.
}
\details{
The complementary graph of an UG is the graph that has the
same set of nodes and an undirected edge connecting \eqn{i}
and \eqn{j} whenever there is not an \eqn{(i,j)} edge in the
original UG.
}
\examples{
## A chordless four-cycle
four <- UG(~ a * b + b * d + d * e + e * a)
four
cmpGraph(four)

}
\references{
Lauritzen, S. (2026). \emph{Graphical models} 2nd ed. Oxford: Oxford
University Press.
}
\seealso{
\code{\link[=UG]{UG()}}, \code{\link[=DAG]{DAG()}}
}
\author{
Giovanni M. Marchetti
}
\keyword{graphs}
\keyword{models}
\keyword{multivariate}
