% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{conComp}
\alias{conComp}
\title{Connectivity components}
\usage{
conComp(amat, method = 1)
}
\arguments{
\item{amat}{A square matrix with dimnames, the adjacency matrix
of an UG.}

\item{method}{An integer 1 or 2 to choose the method used
to find the components. Method 2 is more efficient for large graphs.}
}
\value{
An integer vector representing a partition of the set
of nodes.
}
\description{
Finds the connectivity components of a graph.
}
\examples{
## three connected components
conComp(UG(~ a * c + c * d + e + g * o * u))
## a connected graph
conComp(UG(~ a * b + b * c + c * d + d * a))

}
\references{
#' Lauritzen, S. (2026). \emph{Graphical models} 2nd ed. Oxford: Oxford
University Press.
}
\seealso{
\code{\link[=UG]{UG()}}
}
\author{
Giovanni M. Marchetti
}
\keyword{graphs}
\keyword{models}
\keyword{multivariate}
