% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{correlations}
\alias{correlations}
\title{Marginal and partial correlations}
\usage{
correlations(x, placement = upper.tri)
}
\arguments{
\item{x}{A square symmetric matrix, a covariance matrix,
or a data.frame for \code{n} observations and \code{p} variables.}

\item{placement}{is the function \code{upper.tri} or \verb{lower tri} that specyfies
the placement of the pairwise partial correlations. The default is \code{upper.tri}.}
}
\value{
A square correlation matrix with marginal correlations
and partial correlations with the chosen \code{placement}.
}
\description{
Computes a correlation matrix with ones along the diagonal, marginal
correlations in the lower triangle and partial correlations given all
remaining variables in the upper triangle.
}
\examples{
## See Table 6.1 in Cox & Wermuth (1996)
data(glucose)
correlations(glucose)

correlations(glucose, placement = lower.tri)

}
\references{
Cox, D. R. & Wermuth, N. (1996). \emph{Multivariate dependencies}.
London: Chapman & Hall.
}
\seealso{
\code{\link[=parcor]{parcor()}}, \code{\link[=cor]{cor()}}
}
\author{
Giovanni M. Marchetti
}
\keyword{array}
\keyword{graphs}
\keyword{models}
\keyword{multivariate}
