% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{DAG}
\alias{DAG}
\title{Directed acyclic graphs (DAGs)}
\usage{
DAG(..., order = FALSE)
}
\arguments{
\item{...}{A sequence of model formulae.}

\item{order}{Logical, defaulting to \code{FALSE}. If \code{TRUE} the nodes
of the DAG are permuted according to the topological order.
If \code{FALSE} the nodes are in the order they first appear
in the model formulae (from left to right).}
}
\value{
The adjacency matrix of the DAG, i.e., a square Boolean
matrix of order equal to the number of nodes of the graph and
a one in position \eqn{(i,j)} if there is an arrow from
\eqn{i} to \eqn{j} and zero otherwise. The rownames of the
adjacency matrix are the nodes of the DAG.

If \code{order = TRUE} the adjacency matrix is permuted to have parents
before children. This can always be done (in more than one way)
for DAGs. The resulting adjacency matrix is upper triangular.
}
\description{
A simple way to define a DAG by means of regression model formulae.
}
\details{
The DAG is defined by a sequence of recursive regression models.
Each regression is defined by a model formula.
For each formula the response defines a node of the graph and
the explanatory variables the parents of that node.
If the regressions are not recursive the function returns an
error message.

Some authors prefer the terminology acyclic directed graphs (ADG).
}
\note{
The model formulae may contain interactions,
but they are ignored in the graph.
}
\examples{
# A Markov chain
DAG(y ~ x, x ~ z, z ~ u)

# Another DAG
DAG(y ~ x + z + u, x ~ u, z ~ u)

# A DAG with an isolated node
DAG(v ~ v, y ~ x + z, z ~ w + u)

# There can be repetitions
DAG(y ~ x + u + v, y ~ z, u ~ v + z)

# Interactions are ignored
DAG(y ~ x * z + z * v, x ~ z)

# A cyclic graph returns an error!
\dontrun{
DAG(y ~ x, x ~ z, z ~ y)
}

## The order can be changed
DAG(y ~ z, y ~ x + u + v, u ~ v + z)

## If you want to order the nodes (topological sort of the DAG)
DAG(y ~ z, y ~ x + u + v, u ~ v + z, order = TRUE)

}
\references{
Lauritzen, S. (1996). \emph{Graphical models} 2nd ed.
Oxford: Oxford University Press.
}
\seealso{
\code{\link[=UG]{UG()}}, \code{\link[=topSort]{topSort()}}, \code{\link[=edgematrix]{edgematrix()}}, \code{\link[=fitDag]{fitDag()}}
}
\author{
G. M. Marchetti
}
\keyword{graphs}
\keyword{models}
\keyword{multivariate}
