% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{diagv}
\alias{diagv}
\title{Matrix product with a diagonal matrix}
\usage{
diagv(v, M)
}
\arguments{
\item{v}{A numeric vector specifying the elements on the diagonal of a
matrix.}

\item{M}{A numeric matrix compatible with the product \eqn{D_v M}.}
}
\value{
A matrix \code{N}.
}
\description{
Computes faster the product of a diagonal matrix times a full matrix.
}
\details{
Computes \eqn{N = D_v M} where \eqn{D_v} is diagonal avoiding the
\code{diag} operator.
}
\examples{

v <- 1:1000
M <- matrix(runif(3000), 1000, 3)
dim(diagv(v, M))

}
\seealso{
\code{\link{diag}}
}
\keyword{matrix}
