% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{findPath}
\alias{findPath}
\title{Finding paths}
\usage{
findPath(amat, st, en, path = c())
}
\arguments{
\item{amat}{A square Boolean matrix with dimnames,
the adjacency matrix of a graph.}

\item{st}{An integer, the starting node index.}

\item{en}{An integer, the ending node index.}

\item{path}{A vector of integers, used in recursive calls.
Defaults to an empty vector \code{c()}. It should not be
modified by the user.}
}
\value{
A vector of integers containing the sequence
of nodes forming a path from \code{st} to \code{en}. If no path
is found, it returns \code{NULL}.
In some graphs (like spanning trees) there is only one path
between two nodes.
}
\description{
Finds one path between two nodes of a graph.
}
\examples{
## A (single) path on a spanning tree
findPath(bfsearch(UG(~ a * b * c + b * d + d * e + e * c))$tree, st = 1, en = 5)

}
\references{
Python Software Foundation (2003). Python Patterns
Implementing Graphs.
\url{https://www.python.org/doc/essays/graphs/}.
}
\seealso{
\code{\link[=fundCycles]{fundCycles()}}
}
\author{
Giovanni M. Marchetti, translating the original
Python code (see references).
}
\keyword{graphs}
