% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{findPath}
\alias{findPath}
\title{Finding paths}
\usage{
findPath(amat, st, en, path = c())
}
\arguments{
\item{amat}{a square Boolean matrix with dimnames, the adjacency matrix of a
graph.}

\item{st}{an integer, the starting node.}

\item{en}{an integer, the ending node.}

\item{path}{a vector of integers, used in recursive calls. At the beginning
is \code{NULL}. It should not be modified by the user.}
}
\value{
a vector of integers, the sequence of nodes of a path, starting from
\code{st} to \code{en}. In some graphs (spanning trees) there is only one
path between two nodes.
}
\description{
Finds one path between two nodes of a graph.
}
\note{
This function is not intended to be directly called by the user.
}
\examples{

## A (single) path on a spanning tree
findPath(bfsearch(UG(~ a*b*c + b*d + d*e+ e*c))$tree, st=1, en=5)

}
\references{
Python Softftware Foundation (2003). Python Patterns ---
Implementing Graphs. \url{http://www.python.org/doc/essays/graphs/}.
}
\seealso{
\code{\link{fundCycles}}
}
\author{
Giovanni M. Marchetti, translating the original \pkg{Python} code
(see references).
}
\keyword{graphs}
