% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{fitAncestralGraph}
\alias{fitAncestralGraph}
\title{Fitting of Gaussian Ancestral Graph Models}
\usage{
fitAncestralGraph(amat, S, n, tol = 1e-06)
}
\arguments{
\item{amat}{A square matrix, representing the adjacency matrix of an
ancestral graph.}

\item{S}{A symmetric positive definite matrix with row and col names, the
sample covariance matrix.}

\item{n}{The sample size, a positive integer.}

\item{tol}{A small positive number indicating the tolerance used in
convergence checks. Defaults to 1e-06.}
}
\value{
A list with the following components:
\item{Shat}{The fitted covariance matrix.}
\item{Lhat}{Matrix of the fitted precisions associated with undirected edges
and vertices that do not have an arrowhead pointing at them.}
\item{Bhat}{Matrix of the fitted regression coefficients associated to the directed edges.
Precisely said, \code{Bhat} contains ones on the diagonal and the off-diagonal entry
\eqn{(i,j)}{(i,j)} equals the \emph{negated} MLE of the regression coefficient for
variable \eqn{j}{j} in the regression of variable \eqn{i}{i} on its parents.
Note that this \eqn{(i,j)}{(i,j)} entry in \code{Bhat} corresponds to a directed
edge \code{j -> i}, and thus to a one as \eqn{(j,i)}{(j,i)} entry in the
adjacency matrix.}
\item{Ohat}{Matrix of the error covariances and variances of the residuals between
regression equations associated with bi-directed edges and vertices with an
arrowhead pointing at them.}
\item{dev}{The deviance of the model.}
\item{df}{The degrees of freedom.}
\item{it}{The number of iterations.}
}
\description{
Iterative conditional fitting of Gaussian Ancestral Graph Models.
}
\details{
In the Gaussian case, the models can be parameterized using precision
parameters, regression coefficients, and error covariances (compare
Richardson and Spirtes, 2002, Section 8). This function finds the MLE
\eqn{\hat{\Lambda}}{L} of the precision parameters by fitting a concentration
graph model. The MLE \eqn{\hat{B}}{B} of the regression coefficients and the
MLE \eqn{\hat{\Omega}}{O} of the error covariances are obtained by iterative
conditional fitting (Drton and Richardson, 2003, 2004). The three sets of
parameters are combined to the MLE \eqn{\hat{\Sigma}}{S} of the covariance
matrix by matrix multiplication:
\deqn{\hat{\Sigma} = \hat{B}^{-1}(\hat{\Lambda}+\hat{\Omega})\hat{B}^{-T}}{S = B^(-1) (L+O) B^(-T)}
Note that in Richardson and Spirtes (2002), the matrices \eqn{\Lambda}{L}
and \eqn{\Omega}{O} are defined as submatrices.
}
\examples{
## A covariance matrix
S <- structure(
  c(
    2.93, -1.7, 0.76, -0.06,
    -1.7, 1.64, -0.78, 0.1,
    0.76, -0.78, 1.66, -0.78,
    -0.06, 0.1, -0.78, 0.81
  ),
  .Dim = c(4, 4),
  .Dimnames = list(c("y", "x", "z", "u"), c("y", "x", "z", "u"))
)

## Fit an ancestral graph y -> x <-> z <- u
fitAncestralGraph(makeMG(dg = DAG(x ~ y, z ~ u), bg = UG(~ x * z)), S, n = 100)

## Fit an ancestral graph y <-> x <-> z <-> u
fitAncestralGraph(makeMG(bg = UG(~ y * x + x * z + z * u)), S, n = 100)

## Fit the same graph with fitCovGraph
fitCovGraph(makeMG(bg = UG(~ y * x + x * z + z * u)), S, n = 100)

## Another example for the mathematics marks data
data(marks)
S_marks <- var(marks)
mag1 <- makeMG(bg = UG(~ mechanics * vectors * algebra + algebra * analysis * statistics))
fitAncestralGraph(mag1, S_marks, n = 88)

}
\references{
Drton, M. and Richardson, T. S. (2003). A new algorithm for
maximum likelihood estimation in Gaussian graphical models for marginal
independence. \emph{Proceedings of the Nineteenth Conference on Uncertainty
in Artificial Intelligence}, 184-191.

Drton, M. and Richardson, T. S. (2004). Iterative Conditional Fitting for
Gaussian Ancestral Graph Models. \emph{Proceedings of the 20th Conference on
Uncertainty in Artificial Intelligence}, Department of Statistics, 130-137.

Richardson, T. S. and Spirtes, P. (2002). Ancestral Graph Markov Models.
\emph{Annals of Statistics}, 30(4), 962-1030.
}
\seealso{
\code{\link{fitCovGraph}}, \code{\link{icf}}, \code{\link{makeMG}}, \code{\link{fitDag}}
}
\author{
Mathias Drton
}
\keyword{ancestral.graph}
\keyword{graphs}
\keyword{models}
\keyword{multivariate}
