% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-plot.R
\name{gram_plotOutput}
\alias{gram_plotOutput}
\title{Shiny output binding for a gram plot}
\usage{
gram_plotOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{Output variable to read from.}

\item{width, height}{Valid CSS dimensions.}
}
\value{
A Shiny widget output element.
}
\description{
Shiny output binding for a gram plot
}
\seealso{
Other widgets:
\code{\link[=gram_plot]{gram_plot()}},
\code{\link[=gram_tracingOutput]{gram_tracingOutput()}},
\code{\link[=render_gram_plot]{render_gram_plot()}},
\code{\link[=render_gram_tracing]{render_gram_tracing()}}
}
\concept{widgets}
