% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{hypercube}
\alias{hypercube}
\title{Create a Boolean matrix from all combinations of binary variables}
\usage{
hypercube(n)
}
\arguments{
\item{n}{a positive integer, \eqn{n \in \{1, \dots, 20\}}. If not
the function gives an error message.}
}
\value{
a matrix with \eqn{2^n}{2^n} rows and \eqn{n} columns.
Each row of the matrix is a \code{TRUE}/\code{FALSE} vector
representing the coordinates of the vertices of
a hypercube, ordered in inverse lexicographic order.
}
\description{
Create a Boolean matrix from all combinations of binary variables
}
\examples{
 hypercube(4)

}
\seealso{
\code{\link[base:expand.grid]{base::expand.grid()}}
}
