% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{induced-graphs}
\alias{induced-graphs}
\alias{inducedChainGraph}
\alias{inducedCovGraph}
\alias{inducedConGraph}
\alias{inducedRegGraph}
\alias{inducedDAG}
\title{Graphs induced by marginalization or conditioning}
\usage{
inducedChainGraph(amat, cc = rownames(amat), cond = NULL, type = "LWF")

inducedCovGraph(amat, sel = rownames(amat), cond = NULL)

inducedConGraph(amat, sel = rownames(amat), cond = NULL)

inducedRegGraph(amat, sel = rownames(amat), cond = NULL)

inducedDAG(amat, order, cond = NULL)
}
\arguments{
\item{amat}{a square Boolean matrix, the adjacency matrix of a
directed acyclic graph. The names of rows and of the columns are the
nodes of the DAG.}

\item{cc}{a list of character vectors specifying the chain
components for the chain graph.}

\item{cond}{a character vector representing the variables on which
you want to condition. \code{cond} must be disjoint from \code{sel} and
their union must be a subset of the set of nodes.
The set difference between the set of nodes and the union of
\code{sel} and \code{cond} are the variables over which we marginalize.
\code{cond}  may be the null vector (the default), meaning that you
want to condition on the empty set.}

\item{type}{a string indicating the interpretation of the chain
graph. It can be either
\itemize{
\item \code{"LWF"} (Lauritzen, Wermuth, Frydenberg interpretation),
\item \code{"AMP"} (Andersson, Madigan, Perlman interpretation) or
\item \code{"MRG"} (Multivariate regression graph interpretation).
}}

\item{sel}{a character vector representing a subset of selected
variables. The elements of the vector must be a subset of the names
of the nodes i.e., of \code{rownames(amat)}. By default \code{sel} is the set
of the nodes of the DAG.}

\item{order}{a character vector indicating the ordering of the vertices of a DAG (left to right, past to future).}
}
\value{
\itemize{
\item \code{inducedCovGraph} returns the adjacency matrix of the covariance
graph of the variables in set \code{sel} given the variables in
set \code{cond}, implied by the original directed acyclic graph with
adjacency matrix \code{amat}.
\item \code{inducedConGraph} returns the adjacency matrix of the concentration
graph of the variables in set \code{sel} given the variables in set
\code{cond}, implied by the original directed acyclic graph with
adjacency matrix \code{amat}.
\item \code{inducedRegGraph} returns the adjacency matrix of the multivariate
regression graph of the variables in set \code{sel} given the variables in
set \code{cond}, implied by the original directed acyclic graph with
adjacency matrix \code{amat}.
\item \code{inducedChainGraph} returns the adjacency matrix of the chain graph
for the variables in the chain components \code{cc}, given the variables
in set \code{cond}, with interpretation specified by string
\code{type}, implied by the original directed acyclic graph with
adjacency matrix \code{amat}.
\item \code{inducedDAG} returns the adjacency matrix of the DAG with the
ordering \code{order}, implied by the original directed acyclic graph
with adjacency matrix \code{amat}.
}
}
\description{
Functions to find induced graphs after conditioning on a set of
variables and marginalizing over another set.
}
\details{
Given a directed acyclic graph representing a set of conditional
independencies it is possible to obtain other graphs of conditional
independence implied after marginalizing over and conditionig on
sets of nodes. Such graphs are the covariance graph, the
concentration graph, the multivariate regression graph and the
chain graph with different interpretations (see Cox & Wermuth,
1996, 2004).
}
\note{
If \code{sel = NULL} the functions return the null matrix.
If \code{cond = NULL}, the conditioning set is empty and the functions
\code{inducedConGraph} and \code{inducedCovGraph} return the overall induced
covariance or concentration matrices of the selected variables. I
f you do not specify \code{sel} you cannot specify a non-\code{NULL}
value of \code{cond}.
}
\examples{
if (requireNamespace("ggm", quietly = TRUE)) {
  library(ggm)

  # Define a DAG
  dag <- DAG(a ~ x, c ~ b + d, d ~ x)
  dag
  # Induced covariance graph of a, b, d given the empty set.
  inducedCovGraph(dag, sel = c("a", "b", "d"), cond = NULL)

  # Induced concentration graph of a, b, c given x
  inducedConGraph(dag, sel = c("a", "b", "c"), cond = "x")

  # Overall covariance graph
  inducedCovGraph(dag)

  # Overall concentration graph
  inducedConGraph(dag)

  # Induced covariance graph of x, b, d given c, x.
  inducedCovGraph(dag, sel = c("a", "b", "d"), cond = c("c", "x"))

  # Induced concentration graph of a, x, c given d, b.
  inducedConGraph(dag, sel = c("a", "x", "c"), cond = c("d", "b"))

  # The DAG on p. 198 of Cox & Wermuth (1996)
  dag <- DAG(y1 ~ y2 + y3, y3 ~ y5, y4 ~ y5)

  # Cf. figure 8.7 p. 203 in Cox & Wermuth (1996)
  inducedCovGraph(dag, sel = c("y2", "y3", "y4", "y5"), cond = "y1")
  inducedCovGraph(dag, sel = c("y1", "y2", "y4", "y5"), cond = "y3")
  inducedCovGraph(dag, sel = c("y1", "y2", "y3", "y4"), cond = "y5")

  # Cf. figure 8.8 p. 203 in Cox & Wermuth (1996)
  inducedConGraph(dag, sel = c("y2", "y3", "y4", "y5"), cond = "y1")
  inducedConGraph(dag, sel = c("y1", "y2", "y4", "y5"), cond = "y3")
  inducedConGraph(dag, sel = c("y1", "y2", "y3", "y4"), cond = "y5")

  # Cf. figure 8.9 p. 204 in Cox & Wermuth (1996)
  inducedCovGraph(dag, sel = c("y2", "y3", "y4", "y5"), cond = NULL)
  inducedCovGraph(dag, sel = c("y1", "y2", "y4", "y5"), cond = NULL)
  inducedCovGraph(dag, sel = c("y1", "y2", "y3", "y4"), cond = NULL)

  # Cf. figure 8.10 p. 204 in Cox & Wermuth (1996)
  inducedConGraph(dag, sel = c("y2", "y3", "y4", "y5"), cond = NULL)
  inducedConGraph(dag, sel = c("y1", "y2", "y4", "y5"), cond = NULL)
  inducedConGraph(dag, sel = c("y1", "y2", "y3", "y4"), cond = NULL)

  # An induced regression graph
  dag2 <- DAG(Y ~ X + U, W ~ Z + U)
  inducedRegGraph(dag2, sel = "W", cond = c("Y", "X", "Z"))

  # An induced DAG
  inducedDAG(dag2, order = c("X", "Y", "Z", "W"))

  # An induced multivariate regression graph
  inducedRegGraph(dag2, sel = c("Y", "W"), cond = c("X", "Z"))

  # An induced chain graph with LWF interpretation
  dag3 <- DAG(X ~ W, W ~ Y, U ~ Y + Z)
  cc <- list(c("W", "U"), c("X", "Y", "Z"))
  inducedChainGraph(dag3, cc = cc, type = "LWF")

  # ... with AMP interpretation
  inducedChainGraph(dag3, cc = cc, type = "AMP")

  # ... with multivariate regression interpretation
  cc <- list(c("U"), c("Z", "Y"), c("X", "W"))
  inducedChainGraph(dag3, cc = cc, type = "MRG")
}

}
\references{
Cox, D. R. & Wermuth, N. (1996). \emph{Multivariate dependencies}.
London: Chapman & Hall.

Wermuth, N. & Cox, D.R. (2004). Joint response graphs and separation
induced by triangular systems. \emph{J. R. Statist. Soc. B}, 66, 687-717.
}
\seealso{
\code{\link[=DAG]{DAG()}}, \code{\link[=UG]{UG()}}, \code{\link[=isAcyclic]{isAcyclic()}}
}
\author{
Giovanni M. Marchetti
}
\keyword{conditioning}
\keyword{marginalization}
\keyword{models}
