% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{isADMG}
\alias{isADMG}
\title{Acyclic directed mixed graphs}
\usage{
isADMG(amat)
}
\arguments{
\item{amat}{An adjacency matrix.}
}
\value{
A logical value, \code{TRUE} if it is an ancestral graph and
\code{FALSE} otherwise.
}
\description{
Check if it is an adjacency matrix of an ADMG
}
\details{
Checks if the following conditions must hold: (i) no undirected edge meets
an arrowhead; (ii) no directed cycles;
}
\examples{

	## Examples from Richardson and Spirtes (2002)
	a1 <- makeMG(dg=DAG(a~b, b~d, d~c), bg=UG(~a*c))
	isADMG(a1)    # Not an AG. (a2) p.969
	a2 <- makeMG(dg=DAG(b ~ a, d~c), bg=UG(~a*c+c*b+b*d))           # Fig. 3 (b1) p.969
	isADMG(a2)

}
\references{
Richardson, T. S. and Spirtes, P. (2002). Ancestral Graph Markov
Models. \emph{Annals of Statistics}, 30(4), 962--1030.
}
\seealso{
\code{\link{makeMG}}, \code{\link{isADMG}}
}
\author{
Giovanni M. Marchetti, Mathias Drton
}
\keyword{ancestral}
\keyword{graph}
\keyword{graphs}
\keyword{mixed}
\keyword{models}
\keyword{multivariate}
