% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{MAG}
\alias{MAG}
\title{Maximal ancestral graph}
\usage{
MAG(
  amat,
  M = c(),
  C = c(),
  showmat = TRUE,
  plot = FALSE,
  plotfun = plotGraph,
  ...
)
}
\arguments{
\item{amat}{An adjacency matrix, or a graph object. This can be:
\itemize{
\item a \code{graphNEL} object,
\item an \code{\link[igraph:igraph]{igraph::igraph()}} object, or
\item a character vector of length \eqn{3e}{3e}, where \eqn{e}
is the number of edges. If it is a vector, it must
be a sequence of triples (\code{type}, \code{node1label}, \code{node2label}).

The type of edge can be:
\itemize{
\item \code{"a"} (arrow from \code{node1} to \code{node2}),
\item \code{"b"} (bi-directed arc),
\item \code{"l"} (undirected line), or
\item \code{"*"} (for an isolated node with \code{node1 == node2}).
}
}}

\item{M}{A subset of the node set of \code{amat} that is going to be
marginalized over.}

\item{C}{Another disjoint subset of the node set of \code{amat} that
is going to be conditioned on.}

\item{showmat}{A logical value. \code{TRUE} (by default) to print the
generated matrix.}

\item{plot}{\code{FALSE} (by default). \code{TRUE} to plot the generated graph.}

\item{plotfun}{Function to plot the graph when \code{plot = TRUE}.
Can be \code{plotGraph} (the default) or \code{drawGraph}.}

\item{...}{Further arguments passed to \code{plotfun}.}
}
\value{
A matrix consisting of 4 different integers
representing the \eqn{ij}{ij}-elements:
\itemize{
\item 0 for a missing edge between \eqn{i} and \eqn{j},
\item 1 for an arrow from \eqn{i} to \eqn{j},
\item 10 for a full line between \eqn{i} and \eqn{j}, and
\item 100 for a bi-directed arrow between \eqn{i} and \eqn{j}.
These numbers are added when multiple edges of different types
are present. The matrix is symmetric with respect to full lines
and bi-directed arrows.
}
}
\description{
\code{MAG} generates and plots maximal ancestral graphs after
marginalisation and conditioning.
}
\details{
Support for \code{graphNEL} objects requires the \code{graph}
package from Bioconductor, which is a suggested dependency.
If the package is missing, passing a \code{graphNEL} object will
trigger an informative error.

This function uses the internal functions \code{\link[=AG]{AG()}} and \code{\link[=Max]{Max()}}.
}
\examples{
# The adjacency matrix of a DAG
ex <- matrix(c(
  0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 
  0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
  1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
  0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
  0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
  0, 1, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
  0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
  0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
  0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0,
  0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0,
  0, 0, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0,
  1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
  0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0,
  0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
  1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0,
  0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 0, 0
), 16, 16, byrow = TRUE)
M <- c(3, 5, 6, 15, 16)
C <- c(4, 7)
MAG(ex, M, C, plot = TRUE)

H <- matrix(c(0, 100, 1, 0, 
             100, 0, 100, 0, 
             0, 100, 0, 100, 
             0, 1, 100, 0), 4, 4)
Max(H)

}
\references{
Richardson, T. S. and Spirtes, P. (2002). Ancestral graph Markov
models. \emph{Annals of Statistics}, 30(4), 962-1030.

Sadeghi, K. (2013). Stable mixed graphs. \emph{Bernoulli}, 19(5B),
2330–2358.

Sadeghi, K. and Lauritzen, S.L. (2014). Markov properties for
loopless mixed graphs. \emph{Bernoulli}, 20(2), 676-696.
}
\seealso{
\code{\link[=AG]{AG()}}, \code{\link[=Max]{Max()}}, \code{\link[=MRG]{MRG()}}, \code{\link[=MSG]{MSG()}}
}
\author{
Kayvan Sadeghi
}
\keyword{DAG}
\keyword{ancestral-graph}
\keyword{conditioning}
\keyword{marginalization}
\keyword{maximality}
