% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{marg.param}
\alias{marg.param}
\title{Link function of the marginal log-linear parameterization}
\usage{
marg.param(lev, type)
}
\arguments{
\item{lev}{Integer vector containing the number of levels of each variable.}

\item{type}{A character vector with elements \code{"l"}, \code{"g"},
\code{"c"}, or \code{"r"} indicating the type of logit. The meaning is as
follows:
\itemize{
\item \code{"g"} for global logit
\item \code{"c"} for continuation logit
\item \code{"r"} for the reverse continuation logit
\item \code{"l"} for local logit.
}}
}
\value{
a list  with slots
\itemize{
\item \code{C} Matrix of constrasts (the first \code{sum(lev)-length(r)}
elements are referred to univariate logits)
\item \code{M} Marginalization matrix with elements 0 and 1
\item \code{G} The design matrix for the corresponding log-linear model.
}
}
\description{
Provides the contrast and marginalization matrices
for the marginal parametrization of a probability vector.
}
\details{
See Bartolucci, Colombi and Forcina (2007).

The function assumes that the vector of
probabilities is in inverse lexicographic order.
The interactions are returned in order of dimension,
like e.g., \code{1}, \code{2}, \code{3}, \code{12}, \code{13}, \code{23}, \code{123}.
}
\examples{

marg.param(c(3, 3), c("l", "g"))

}
\references{
Bartolucci, F., Colombi, R. and Forcina, A. (2007).
An extended class of marginal link functions for modelling
contingency tables by equality and inequality constraints.
\emph{Statistica Sinica},  17, 691-711.
}
\seealso{
\code{\link[=mat.mlogit]{mat.mlogit()}}.
}
\author{
Francesco Bartolucci, Antonio Forcina. Giovanni M.
Marchetti translated the Matlab function by Bartolucci and Forcina.
}
\keyword{logistic}
\keyword{models}
\keyword{ordinal.}
