% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{marg.param}
\alias{marg.param}
\title{Link function of marginal log-linear parameterization}
\usage{
marg.param(lev, type)
}
\arguments{
\item{lev}{Integer vector containing the number of levels of each variable.}

\item{type}{A character vector with elements \code{"l"}, \code{"g"},
\code{"c"}, or \code{"r"} indicating the type of logit. The meaning is as
follows: \code{"g"} for global, \code{"c"} for continuation, \code{"r"} for
reverse continuation and \code{"l"} for local.}
}
\value{
\item{C}{Matrix of constrasts (the first \code{sum(lev)-length(r)}
elements are referred to univariate logits)} \item{M}{Marginalization matrix
with elements 0 and 1.} \item{G}{Corresponding design matrix for the
corresponding log-linear model.}
}
\description{
Provides the contrast and marginalization matrices for the marginal
parametrization of a probability vector.
}
\details{
See Bartolucci, Colombi and Forcina (2007).
}
\note{
Assumes that the vector of probabilities is in inv lex order.  The
interactions are returned in order of dimension, like e.g., 1, 2, 3, 12, 13,
23, 123.
}
\examples{
    
marg.param(c(3,3), c("l", "g"))

}
\references{
Bartolucci, F., Colombi, R. and Forcina, A. (2007). An extended
class of marginal link functions for modelling contingency tables by
equality and inequality constraints. Statist. Sinica 17, 691-711.
}
\seealso{
\code{\link{mat.mlogit}}
}
\author{
Francesco Bartolucci, Antonio Forcina, Giovanni M. Marchetti
}
\keyword{logistic}
\keyword{models}
\keyword{ordinal}
