% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{MarkEqRcg}
\alias{MarkEqRcg}
\title{Markov equivalence for regression chain graphs}
\usage{
MarkEqRcg(amat, bmat)
}
\arguments{
\item{amat}{An adjacency matrix of an RCG, or a graph that can be a
\code{graphNEL} object, an \link[igraph:igraph]{igraph::igraph} object, or a character
vector of length \verb{3e} (where \code{e} is the number of edges).
If a vector, it must be a sequence of triples
(\code{type}, \code{node1label}, \code{node2label}). The \code{type} of edge can be
\itemize{
\item \code{"a"} (arrows from node1 to node2),
\item \code{"b"} (arcs), or
\item \code{"l"} (lines).
}}

\item{bmat}{The same as \code{amat}.}
}
\value{
A character string: either \code{"Markov Equivalent"} or \code{"NOT Markov Equivalent"}.
}
\description{
\code{MarkEqRcg} determines whether two regression chain graphs (RCGs) or
subclasses of RCGs are Markov equivalent. The function checks whether the two graphs have the same skeleton and "unshielded colliders".
}
\examples{
H1 <- matrix(c(
  0, 100, 0, 0, 0, 
  100, 0, 100, 0, 0, 
  0, 100, 0, 0, 0, 
  1, 0, 0, 0, 100, 
  0, 0, 1, 100, 0
), 5, 5)

H2 <- matrix(c(
  0, 0, 0, 0, 0, 
  1, 0, 1, 0, 0, 
  0, 0, 0, 0, 0, 
  1, 0, 0, 0, 100, 
  0, 0, 1, 100, 0
), 5, 5)

H3 <- matrix(c(
  0, 0, 0, 0, 0, 
  1, 0, 1, 0, 0, 
  0, 0, 0, 0, 0, 
  1, 0, 0, 0, 0, 
  0, 0, 1, 1, 0
), 5, 5)

MarkEqRcg(H1, H2)
MarkEqRcg(H1, H3)
MarkEqRcg(H2, H3)

}
\references{
Wermuth, N. and Sadeghi, K. (2012). Sequences of regressions and
their independences. \emph{Test}, 21:215-252.
}
\seealso{
\code{\link[=msep]{msep()}}, \code{\link[=grMAT]{grMAT()}}
}
\author{
Kayvan Sadeghi
}
\keyword{Markov-equivalence}
\keyword{regression-chain-graph}
\keyword{undirected-graph}
