% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggm_latest-package.R
\docType{data}
\name{marks}
\alias{marks}
\title{Mathematics marks}
\format{
A data frame with 88 observations on the following 5 variables.
\describe{ \item{mechanics}{a numeric vector, mark in Mechanics}
\item{vectors}{a numeric vector, mark in Vectors} \item{algebra}{a numeric
vector, mark in Algebra} \item{analysis}{a numeric vector, mark in Analysis}
\item{statistics}{a numeric vector, mark in Statistics } }
}
\source{
Mardia, K.V., Kent, J.T. and Bibby, (1979). \emph{Multivariate
analysis}. London: Academic Press.
}
\description{
Examination marks of 88 students in five subjects.
}
\details{
Mechanics and Vectors were closed book examinations. Algebra, Analysis and
Statistics were open book examinations.
}
\examples{

data(marks)
pairs(marks)

}
\references{
Whittaker, J. (1990). \emph{Graphical models in applied
multivariate statistics}. Chichester: Wiley.
}
\keyword{datasets}
