% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{Max}
\alias{Max}
\title{Maximization for graphs}
\usage{
Max(amat)
}
\arguments{
\item{amat}{An adjacency matrix, or a graph that can be a \code{graphNEL}
or an \code{\link[igraph:igraph]{igraph::igraph()}} object or a vector of length \eqn{3e},
where \eqn{e} is the number of edges of the graph, that is
a sequence of triples (\code{type}, \code{node1label}, \code{node2label}).
The type of edge can be
\itemize{
\item \code{"a"} (arrows from \code{node1} to \code{node2}),
\item \code{"b"} (arcs), and
\item \code{"l"} (lines).
}}
}
\value{
A matrix that consists 4 different integers as
an \eqn{ij}-element:
\itemize{
\item 0 for a missing edge between \eqn{i} and \eqn{j},
\item 1 for an arrow from \eqn{i} to \eqn{j},
\item 10 for a full line between \eqn{i} and \eqn{j}, and
\item 100 for a bi-directed arrow between \eqn{i} and \eqn{j}.
These numbers are added to be associated with multiple edges of
different types. The matrix is symmetric w.r.t full lines and
bi-directed arrows.
}
}
\description{
\code{Max} generates a maximal graph that induces the same independence
model from a non-maximal graph.
}
\details{
\code{Max} looks for non-adjacent pais of nodes that are connected by
primitive inducing paths, and connect such pairs by an
appropriate edge.
}
\examples{

H <- matrix(c(
  0, 100, 1, 0,
  100, 0, 100, 0,
  0, 100, 0, 100,
  0, 1, 100, 0
), 4, 4)
Max(H)

}
\references{
Richardson, T.S. and Spirtes, P. (2002). Ancestral graph Markov
models. \emph{Annals of Statistics}, 30(4), 962-1030.

Sadeghi, K. and Lauritzen, S.L. (2014). Markov properties for
loopless mixed graphs. \emph{Bernoulli}, 20(2), 676-696.
}
\seealso{
\code{\link[=MAG]{MAG()}}, \code{\link[=MRG]{MRG()}}, \code{\link[=msep]{msep()}}, \code{\link[=MSG]{MSG()}}
}
\author{
Kayvan Sadeghi
}
\keyword{graphs}
\keyword{loopless}
\keyword{m-separation}
\keyword{maximality}
\keyword{mixed-graph}
