% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{null}
\alias{null}
\title{Null space of a matrix}
\usage{
null(M)
}
\arguments{
\item{M}{A matrix.}
}
\value{
The matrix \code{N} with the basis for the null space, or an empty
vector if the matrix \code{M} is square and of maximal rank.
}
\description{
Given a matrix \code{M} find a matrix \code{N} such that \eqn{N^T M} is
zero.
}
\examples{

 null(c(1,1,1))

}
\seealso{
\code{\link[MASS]{Null}}, ~~~
}
\keyword{matrix}
