% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{parcor}
\alias{parcor}
\title{Partial correlations}
\usage{
parcor(S)
}
\arguments{
\item{S}{a symmetric positive definite matrix, representing a covariance
matrix.}
}
\value{
A symmetric matrix with ones along the diagonal and in position
\eqn{(r,s)} the partial correlation between variables \eqn{r} and \eqn{s}
given all the remaining variables.
}
\description{
Finds the matrix of the partial correlations between pairs of variables
given the rest.
}
\details{
The algorithm computes \eqn{- \omega_{rs}/(\omega_{rr} \omega_{ss})^{1/2}}{-w[r,s] /(w[r,r] * w[s,s])^(1/2)}
where \eqn{\omega_{rs}}{w[r,s]} are concentrations,
i.e. elements  of the inverse covariance matrix.
}
\examples{

# Partial correlations for the mathematics marks data
data(marks)
S <- var(marks)
parcor(S)

}
\references{
Cox, D. R. & Wermuth, N. (1996). *Multivariate dependencies].
London: Chapman & Hall.
}
\seealso{
\code{\link[stats:var]{stats::var()}}, \code{\link[stats:cor]{stats::cor()}}, \code{\link[=correlations]{correlations()}}
}
\author{
Giovanni M. Marchetti
}
\keyword{array}
\keyword{graphs}
\keyword{models}
\keyword{multivariate}
