% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{parcor}
\alias{parcor}
\title{Partial correlations}
\usage{
parcor(S)
}
\arguments{
\item{S}{a symmetric positive definite matrix, representing a covariance
matrix.}
}
\value{
A symmetric matrix with ones along the diagonal and in position
\eqn{(r,s)} the partial correlation between variables \eqn{r} and \eqn{s}
given all the remaining variables.
}
\description{
Finds the matrix of the partial correlations between pairs of variables
given the rest.
}
\details{
The algorithm computes \eqn{- \sigma^{rs}/(\sigma^{rr} \sigma^{ss})^{1/2}}
where the \eqn{\sigma^{rs}} are concentrations, i.e. elements of the inverse
covariance matrix.
}
\examples{

### Partial correlations for the mathematics marks data
data(marks)
S <- var(marks)
parcor(S)

}
\references{
Cox, D. R. \& Wermuth, N. (1996). \emph{Multivariate
dependencies}. London: Chapman \& Hall.
}
\seealso{
\code{\link{var}}, \code{\link{cor}}, \code{\link{correlations}}
}
\author{
Giovanni M. Marchetti
}
\keyword{array}
\keyword{graphs}
\keyword{models}
\keyword{multivariate}
