% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{powerset}
\alias{powerset}
\title{Power set}
\usage{
powerset(set, sort = TRUE, nonempty = TRUE)
}
\arguments{
\item{set}{A numeric or character vector.}

\item{sort}{Logical value. If \code{TRUE} (the default) the subsets are sorted according to set dimension. Otherwise the sets are in inverse lexicographical order.}

\item{nonempty}{Logical value. If \code{TRUE} (the default)  the empty set is omitted.}
}
\value{
A list of all subsets of \code{set}.
}
\description{
Finds the power set, that is the list of all subsets of a set.
}
\examples{

powerset(c("A", "B", "C"), nonempty = FALSE)
powerset(1:3, sort = FALSE, nonempty = TRUE)

}
\seealso{
\code{\link[=hypercube]{hypercube()}}, \code{\link[utils:combn]{utils::combn()}}
}
\author{
Giovanni M. Marchetti
}
\keyword{powerset}
\keyword{set}
