% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{powerset}
\alias{powerset}
\title{Power set}
\usage{
powerset(set, sort = TRUE, nonempty = TRUE)
}
\arguments{
\item{set}{A numeric or character vector.}

\item{sort}{Logical value. If \code{TRUE} the subsets are sorted according
to dimension.  Default is \code{TRUE}.}

\item{nonempty}{Logical value. If \code{TRUE} the empty set is omitted.
Default is \code{TRUE}.}
}
\value{
A list of all subsets of \code{set}.
}
\description{
Finds the list of all subsets of a set.
}
\details{
If \code{sort == FALSE} the sets are in inverse lexicographical order.
}
\examples{

powerset(c("A", "B", "C"), nonempty = FALSE)
powerset(1:3, sort = FALSE, nonempty = TRUE)

}
\author{
Giovanni M. Marchetti
}
\keyword{sets}
