% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rcorr}
\alias{rcorr}
\title{Random correlation matrix}
\usage{
rcorr(d)
}
\arguments{
\item{d}{an integer > 0, the order of the correlation matrix.}
}
\value{
a correlation matrix of order \code{d}.
}
\description{
Generates a random correlation matrix with the method of Marsaglia and Olkin
(1984).
}
\details{
The algorithm uses \code{\link{rsphere}} to generate \eqn{d} vectors on a
sphere in \eqn{d}-space. If \eqn{Z} is a matrix with such vectors as rows,
then the random correlation matrix is \eqn{ZZ'}.
}
\examples{

## A random correlation matrix of order 3
rcorr(3)
## A random correlation matrix of order 5
rcorr(5)

}
\references{
Marshall, G.\& Olkin, I. (1984).Generating correlation matrices.
\emph{SIAM J. Sci. Stat. Comput.}, 5, 2, 470--475.
}
\seealso{
\code{\link{rsphere}}
}
\author{
Giovanni M. Marchetti
}
\keyword{distribution}
\keyword{multivariate}
