% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracing-svg.R
\name{render_gram_tracing}
\alias{render_gram_tracing}
\title{Shiny render function for a tracing}
\usage{
render_gram_tracing(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{Expression that produces a \code{\link[=gram_tracing]{gram_tracing()}} widget.}

\item{env}{Environment in which to evaluate \code{expr}.}

\item{quoted}{Whether \code{expr} is quoted.}
}
\value{
A Shiny render function.
}
\description{
Shiny render function for a tracing
}
\seealso{
Other widgets:
\code{\link[=gram_plot]{gram_plot()}},
\code{\link[=gram_plotOutput]{gram_plotOutput()}},
\code{\link[=gram_tracingOutput]{gram_tracingOutput()}},
\code{\link[=render_gram_plot]{render_gram_plot()}}
}
\concept{widgets}
\keyword{internal}
