% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{RepMarUG}
\alias{RepMarUG}
\title{Representational Markov equivalence to undirected graphs.}
\usage{
RepMarUG(amat)
}
\arguments{
\item{amat}{An adjacency matrix, or a graph that can be a
\code{graphNEL} or an \code{\link[igraph:igraph]{igraph::igraph()}} object or a vector of
length \eqn{3e}, where \eqn{e} is the number of edges of the graph,
that is a sequence of triples (\code{type}, \code{node1label}, \code{node2label}).
The type of edge can be
\itemize{
\item \code{"a"} (arrows from \code{node1} to \code{node2},
\item \code{"b"} (arcs), and
\item \code{"l"} (lines).
}}
}
\value{
A list with two components: \code{verify} and \code{amat}.
\itemize{
\item \code{verify} is a logical value, \code{TRUE} if there is a representational
Markov equivalence and \code{FALSE} otherwise.
\item \code{amat} is either \code{NA} if \code{verify} ia \code{FALSE} or is the
adjacency matrix of the generated graph, if \code{verify} is \code{TRUE}.
In this case \code{amat} represent a mixed graph consisting of 4
different integers as an \eqn{ij}-element:
\item 0 for a missing edge between \eqn{i} and \eqn{j},
\item 1 for an arrow from \eqn{i} to \eqn{j},
\item 10 for a full line between \eqn{i} and \eqn{j}, and
\item 100 for a bi-directed arrow between \eqn{i} and#' \eqn{j}.
These numbers are added to be associated with multiple edges of
different types. The matrix is symmetric w.r.t full lines
and bi-directed arrows.
}
}
\description{
\code{RepMarUG} determines whether a given maximal ancestral graph can be
Markov equivalent to an undirected graph, and if that is the case,
it finds an undirected graph that is Markov equivalent to
the given graph.
}
\details{
\code{RepMarUG} looks for presence of an unshielded collider
V-configuration in graph.
}
\examples{

H <- matrix(c(0, 10, 0, 0, 
              10, 0, 0, 0, 
               0, 1, 0, 100, 
               0, 0, 100, 0), 4, 4)
RepMarUG(H)

}
\references{
Sadeghi, K. (2011). Markov equivalences for subclasses of
loopless mixed graphs. \emph{Arxiv}, https://arxiv.org/abs/1110.4539.
}
\seealso{
\code{\link[=MarkEqMag]{MarkEqMag()}}, \code{\link[=MarkEqRcg]{MarkEqRcg()}}, \code{\link[=RepMarBG]{RepMarBG()}},
\code{\link[=RepMarDAG]{RepMarDAG()}}
}
\author{
Kayvan Sadeghi
}
\keyword{Markov-equivalence}
\keyword{Representational}
\keyword{undirected-graph}
