% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{rsphere}
\alias{rsphere}
\title{Random vectors on a sphere}
\usage{
rsphere(n, d)
}
\arguments{
\item{n}{an integer, the sample size.}

\item{d}{an integer, the dimension of the space. For example, a circle is
defined in 2D-space, a sphere in 3D-space.}
}
\value{
a matrix of \code{n} rows and \code{d} columns.
}
\description{
Generates a sample of points uniformly distributed on the surface of a
sphere in d-space.
}
\details{
The algorithm is based on normalizing to length 1 each d-vector of a sample
from a multivariate normal \eqn{N(0, I)}.
}
\examples{

## 100 points on circle
z <- rsphere(100,2)
plot(z)

## 100 points on a sphere
z <- rsphere(100, 3)
pairs(z)

}
\seealso{
\code{\link{rnorm}}, \code{\link{rcorr}}
}
\author{
Giovanni M. Marchetti
}
\keyword{distribution}
\keyword{multivariate}
