% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggm_latest-package.R
\docType{data}
\name{stress}
\alias{stress}
\title{Stress}
\format{
A \eqn{4 \times 4} covariance matrix for the following variables.
\describe{ \item{Y}{} \item{V}{} \item{X}{} \item{U}{} }
}
\description{
Stress data
}
\details{
See Cox and Wermuth (1996).
}
\examples{

data(stress)
G = UG(~ Y*X + X*V + V*U + U*Y)
fitConGraph(G, stress, 100)

}
\references{
Cox, D. R. \& Wermuth, N. (1996). \emph{Multivariate
dependencies}. London: Chapman \& Hall.

Slangen K., Kleemann P.P and Krohne H.W. (1993). Coping with surgical
stress. In: Krohne H. W. (ed.).  \emph{Attention and avoidance: Strategies
in coping with aversiveness}. New York, Heidelberg: Springer, 321-346.
}
\keyword{datasets}
