% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggm-datasets.R
\docType{data}
\name{surdata}
\alias{surdata}
\title{A simulated data set}
\format{
A data frame with 600 observations on the following 4 variables.
\describe{
\item{A}{a numeric response vector}
\item{B}{a numeric response vector}
\item{X}{a numeric vector}
\item{Z}{a numeric vector with codes \code{1} and \code{-1} for a binary variable}
}
}
\usage{
surdata
}
\description{
Simulated data following a seemingly unrelated regression model.
}
\examples{

data(surdata)
pairs(surdata)

}
\keyword{datasets}
