% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracing-svg.R
\name{tracing_spec}
\alias{tracing_spec}
\title{Compile a tracing to SVG and a timeline spec}
\usage{
tracing_spec(x)
}
\arguments{
\item{x}{A \code{Tracing}.}
}
\value{
A list with \code{version}, \code{width}, \code{height}, \code{svg}, and \code{ops}.
}
\description{
Compile a tracing to SVG and a timeline spec
}
\seealso{
Other tracing:
\code{\link[=at]{at()}},
\code{\link[=ggm_tracing]{ggm_tracing()}},
\code{\link[=tracing]{tracing()}}
}
\concept{tracing}
