% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{UG}
\alias{UG}
\title{Defining an undirected graph (UG)}
\usage{
UG(f)
}
\arguments{
\item{f}{a single model formula without response}
}
\value{
a Boolean matrix with dimnames, the adjacency
matrix of the undirected graph.
}
\description{
A simple way to define an undirected graph by means
of a single model formula.
}
\details{
The undirected graph \eqn{\mathcal{G} = (V, E)} is defined by a set of nodes \eqn{V} and a set of pairs \eqn{E}.
The set of pairs is defined by the set of
interactions in the formula. Interactions define complete
subgraphs (not necessarily maximal) of the UG. The best way
is to specify interactions that match the cliques of the
undirected graph. This is the standard way to
define graphical models for contingency tables.
Remember that some hierarchical models are not graphical,
but they nevertheless imply the same graph.

The function returns the adjacency matrix of the
graph, i.e.,  a symmetric Boolean matrix of order equal
to the number of nodes of the graph and a 1 (or a 10) in
position \eqn{(ij)} and \eqn{ji} if there is an edge from
\eqn{j} to \eqn{i} and 0 otherwise. By default, this matrix
has zeros along the main diagonal. The dimnames of
the adjacency matrix are the nodes of the UG.
}
\examples{
# X independent of Y given Z
UG(~ X * Z + Y * Z)

# The saturated model
UG(~ X * Y * Z)

# The model without three-way interactions has the same graph
UG(~ (X + Y + Z)^2)
UG(~ X * Y + Y * Z + Z * X)

# Butterfly model defined from the cliques (see 
# dataset `ggm::marks`)
UG(~ mec * vec * alg + alg * ana * sta)

# Some isolated nodes
UG(~ x * y * z + a + b)

}
\references{
Lauritzen, S. (1996). \emph{Graphical models}, 2nd ed.
Oxford: Oxford University Press.
}
\seealso{
\code{\link[=fitConGraph]{fitConGraph()}}, \code{\link[=fitCovGraph]{fitCovGraph()}}, \code{\link[=DAG]{DAG()}}
}
\author{
Giovanni M. Marchetti
}
\keyword{models}
\keyword{multivariate}
\keyword{undirected-graph}
